The microorganisms Package

Welcome to the microorganisms package.

The microorganisms package is a free and open-source R package with zero dependencies to work with recent data about the microbial taxonomy.

Source

To cite 'microorganisms' in publications use:

Berends MS, Luz CF, Friedrich AW, Sinha BNM, Albers CJ, Glasner C (2022). "AMR: An R Package for Working with Antimicrobial Resistance Data." Journal of Statistical Software, 104(3), 1-31. doi:10.18637/jss.v104.i03 .

A BibTeX entry for LaTeX users is:

@Manual{,
  title = {microorganisms: Microbial Taxonomy Reference Data},
  author = {Matthijs S. Berends and Christian F. Luz and Dennis Souverein and Erwin E. A. Hassing},
  year = {2023},
  note = {https://msberends.github.io/microorganisms/, https://github.com/msberends/microorganisms},
}

See also

Author

Maintainer: Matthijs S. Berends m.s.berends@umcg.nl (ORCID)

Authors:

Christian F. Luz (ORCID) [contributor]

Dennis Souverein (ORCID) [contributor]

Erwin E. A. Hassing [contributor]